+++ src/components/Breadcrumb/index.ts
@@ -1,2 +1,1 @@
 import "./Breadcrumb";
-export type { CandyBreadcrumbProps } from "./Breadcrumb";