+++ src/components/Button/ButtonStyle.ts
@@ -2,11 +2,20 @@ import { css } from "lit";
 
 export default css`
   .button-container {
-    border: color: transparent;
+    border-color: transparent;
     font-weight: 500;
     text-decoration: none;
     border-style: none;
     font-family: inherit;
+    display: flex;
+    gap: 0.5rem;
+    align-items: center;
+    justify-content: center;
+    min-height: 2.5rem;
+  }
+
+  .button-container:hover {
+    cursor: pointer;
   }
 
   .primary {
@@ -14,7 +23,7 @@ export default css`
     background-color: rgb(37 99 235);
   }
 
-   .primary:hover {
+  .primary:hover {
     background-color: rgb(29 78 216);
   }
 
@@ -45,7 +54,7 @@ export default css`
   }
 
   .sm {
-    border-radius: .25rem;
+    border-radius: 0.25rem;
     font-size: 0.75rem;
     line-height: 1rem;
     padding: 1rem;
@@ -53,16 +62,22 @@ export default css`
   }
 
   .md {
-    border-radius: .375rem;
+    border-radius: 0.375rem;
     font-size: 0.875rem;
     line-height: 1.25rem;
-    padding: .5rem 1rem;
+    padding: 0.5rem 1rem;
   }
 
   .xl {
-    border-radius: .375rem;
+    border-radius: 0.375rem;
     font-size: 1rem;
     line-height: 1.5rem;
     padding: 0.75rem 1.5rem;
+  }
+
+  ::slotted(*) {
+    text-decoration: none;
+    color: inherit;
+    transition: 0.2s;
   }
 `;