+++ src/components/Sidebar/Element/Element.ts
@@ -1,8 +1,15 @@
-import { LitElement, html } from "lit";
+import { html } from "lit";
 import { customElement, property } from "lit/decorators.js";
-import ElementStyle from "./ElementStyle";
 
-export type CandySidebarElementProps = {
+import {
+  PopoverInComponentHandler,
+  PopoverInComponentHandlerProps,
+} from "../../Shared/PopoverInComponentHandler";
+import ElementStyle from "./ElementStyle";
+import { verticalMenuIcon } from "../../Shared/vertical-menu-icon";
+import "../../Popover";
+
+export type CandySidebarElementProps = PopoverInComponentHandlerProps & {
   label: string;
   active: boolean;
   disabled: boolean;
@@ -10,8 +17,8 @@ export type CandySidebarElementProps = {
 };
 
 @customElement("candy-sidebar-element")
-export class CandySidebarElement extends LitElement {
-  static styles = ElementStyle;
+export class CandySidebarElement extends PopoverInComponentHandler {
+  static styles = [PopoverInComponentHandler.styles, ElementStyle];
 
   @property({ type: String })
   label = "";
@@ -25,8 +32,36 @@ export class CandySidebarElement extends LitElement {
   @property({ type: Boolean })
   collapsed = false;
 
+  private handleClick(): void {
+    if (this.collapsed) {
+      const reopen = new CustomEvent("onReopen", {
+        bubbles: true,
+        composed: true,
+      });
+      this.dispatchEvent(reopen);
+    }
+  }
+
   render() {
     const styleClass = this.active ? "bg-gray text-blue" : "text-black";
+
+    const renderOptionsIcons =
+      this.minimizeOptions && this.countOptionsSlotAmount() >= 2
+        ? html` <div class="end-icons">
+            <candy-popover class="options-container" side=${this.position}>
+              <div class="vertical-menu-icon" data-testid="sidebar-element-options-icon">
+                ${verticalMenuIcon}
+              </div>
+              <div slot="content" class="options-container">
+                <slot name="options"></slot>
+              </div>
+            </candy-popover>
+          </div>`
+        : html`<div class="end-icons">
+            <div class="options-container">
+              <slot name="options"></slot>
+            </div>
+          </div> `;
 
     return html`
       <button
@@ -36,9 +71,11 @@ export class CandySidebarElement extends LitElement {
           ? "element-container-extended"
           : null}"
         ?disabled="${this.disabled}"
+        @click="${this.handleClick}"
       >
         <slot name="icon"></slot>
-        <p>${!this.collapsed ? this.label : null}</p>
+        <p part="sidebar-element-text">${!this.collapsed ? this.label : null}</p>
+        ${!this.collapsed ? renderOptionsIcons : null}
       </button>
     `;
   }