+++ src/stories/Breadcrumb/renderBreadcrumb.ts
@@ -1,0 +1,14 @@
+import { TemplateResult, html } from "lit";
+import { CandyBreadcrumbElementProps } from "../../components/Breadcrumb/Element";
+import { CandyBreadcrumbControl } from "./BreadcrumbMeta";
+
+export const renderBreadcrumb = (args: CandyBreadcrumbControl) => {
+  return html`<candy-breadcrumb>
+    ${args.elements.map((props: CandyBreadcrumbElementProps) => renderBreadcrumbElement(props))}
+  </candy-breadcrumb>`;
+};
+
+const renderBreadcrumbElement = (props: CandyBreadcrumbElementProps): TemplateResult => {
+  return html`<candy-breadcrumb-element label=${props.label} ?isLast=${props.isLast}>
+  </candy-breadcrumb-element>`;
+};