+++ CHANGELOG.md
@@ -1,3 +1,58 @@
+# [v1.16.0](https://github.com/Candy-Doc/candy-doc-components-library/compare/v1.15.0...v1.16.0) (2023-07-17)
+
+## ♻ Code refactored
+- [`c4b84e8`](https://github.com/Candy-Doc/candy-doc-components-library/commit/c4b84e8)  refactor: add new child component for breadcrumb (#65) (Issues: [`#65`](https://github.com/Candy-Doc/candy-doc-components-library/issues/65))
+
+# [v1.15.0](https://github.com/Candy-Doc/candy-doc-components-library/compare/v1.14.1...v1.15.0) (2023-07-11)
+
+## ✨ New minor Features
+- [`e0f3a3a`](https://github.com/Candy-Doc/candy-doc-components-library/commit/e0f3a3a)  feat: implement popover inside accordions and sidebar elements (#69) (Issues: [`#69`](https://github.com/Candy-Doc/candy-doc-components-library/issues/69))
+
+# [v1.14.1](https://github.com/Candy-Doc/candy-doc-components-library/compare/v1.14.0...v1.14.1) (2023-07-05)
+
+## 🐛 Bug Fixes
+- [`5a24a98`](https://github.com/Candy-Doc/candy-doc-components-library/commit/5a24a98)   fix: css part selector for accordion and sidebar-element (#71) (Issues: [`#71`](https://github.com/Candy-Doc/candy-doc-components-library/issues/71))
+
+# [v1.14.0](https://github.com/Candy-Doc/candy-doc-components-library/compare/v1.13.2...v1.14.0) (2023-07-04)
+
+## ✨ New minor Features
+- [`c613d22`](https://github.com/Candy-Doc/candy-doc-components-library/commit/c613d22)  feat: popover-component (#62) (Issues: [`#62`](https://github.com/Candy-Doc/candy-doc-components-library/issues/62))
+
+# [v1.13.2](https://github.com/Candy-Doc/candy-doc-components-library/compare/v1.13.1...v1.13.2) (2023-06-28)
+
+## 🐛 Bug Fixes
+- [`4b01978`](https://github.com/Candy-Doc/candy-doc-components-library/commit/4b01978)  fix: css file to ts (#68) (Issues: [`#68`](https://github.com/Candy-Doc/candy-doc-components-library/issues/68))
+
+# [v1.13.1](https://github.com/Candy-Doc/candy-doc-components-library/compare/v1.13.0...v1.13.1) (2023-06-26)
+
+## 🐛 Bug Fixes
+- [`e92ad7c`](https://github.com/Candy-Doc/candy-doc-components-library/commit/e92ad7c)  fix(floating button): fix style (#66) (Issues: [`#66`](https://github.com/Candy-Doc/candy-doc-components-library/issues/66))
+
+# [v1.13.0](https://github.com/Candy-Doc/candy-doc-components-library/compare/v1.12.1...v1.13.0) (2023-06-22)
+
+## ✨ New minor Features
+- [`a903cff`](https://github.com/Candy-Doc/candy-doc-components-library/commit/a903cff)  feat(floating button): add a floating button (#63) (Issues: [`#63`](https://github.com/Candy-Doc/candy-doc-components-library/issues/63))
+
+# [v1.12.1](https://github.com/Candy-Doc/candy-doc-components-library/compare/v1.12.0...v1.12.1) (2023-06-20)
+
+## 🐛 Bug Fixes
+- [`18f4d10`](https://github.com/Candy-Doc/candy-doc-components-library/commit/18f4d10)  fix: reopen sidebar when clicking children (#61) (Issues: [`#61`](https://github.com/Candy-Doc/candy-doc-components-library/issues/61))
+
+# [v1.12.0](https://github.com/Candy-Doc/candy-doc-components-library/compare/v1.11.0...v1.12.0) (2023-06-14)
+
+## ✨ New minor Features
+- [`bcc565f`](https://github.com/Candy-Doc/candy-doc-components-library/commit/bcc565f)  feat: add sidebar element&#x27;s options icons (#59) (Issues: [`#59`](https://github.com/Candy-Doc/candy-doc-components-library/issues/59))
+
+# [v1.11.0](https://github.com/Candy-Doc/candy-doc-components-library/compare/v1.10.1...v1.11.0) (2023-06-14)
+
+## 💄 UI and style files added or updated
+- [`cb8ed08`](https://github.com/Candy-Doc/candy-doc-components-library/commit/cb8ed08)  style: Add icon slot in button component (#58) (Issues: [`#58`](https://github.com/Candy-Doc/candy-doc-components-library/issues/58))
+
+# [v1.10.1](https://github.com/Candy-Doc/candy-doc-components-library/compare/v1.10.0...v1.10.1) (2023-06-12)
+
+## 🐛 Bug Fixes
+- [`727b585`](https://github.com/Candy-Doc/candy-doc-components-library/commit/727b585)  style: add missing cursor pointer css on hover (#54) (Issues: [`#54`](https://github.com/Candy-Doc/candy-doc-components-library/issues/54))
+
 # [v1.10.0](https://github.com/Candy-Doc/candy-doc-components-library/compare/v1.9.0...v1.10.0) (2023-06-12)
 
 ## ✨ New minor Features